+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Unreleased
-- 
+
+-
+
+# v2.5.0
+
+- [changed] Import context from `golang.org/x/net` for 1.6 compatibility
+
+### Cloud Messaging
+
+- [added] Added the `messaging` package for sending Firebase notifications
+  and managing topic subscriptions.
+
+### Authentication
+
+- [added] A new [`VerifyIDTokenAndCheckRevoked()`](https://godoc.org/firebase.google.com/go/auth#Client.VerifyIDToken)
+  function has been added to check for revoked ID tokens.
+- [added] A new [`RevokeRefreshTokens()`](https://godoc.org/firebase.google.com/go/auth#Client.RevokeRefreshTokens)
+  function has been added to invalidate all refresh tokens issued to a user.
+- [added] A new property `TokensValidAfterMillis` has been added to the
+  ['UserRecord'](https://godoc.org/firebase.google.com/go/auth#UserRecord)
+  type, which stores the time of the revocation truncated to 1 second accuracy.
 
 # v2.4.0